backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-rest-framework>=0.1.0",
    "django>=5.2",
    "python-dotenv>=1.1.0",
    "pyjwt>=2.10.1",
    "requests>=2.32.3",
    "cryptography>=44.0.3",
    "pytz>=2025.2",
]
